+++ src/lib.rs
@@ -1,2 +1,3 @@
+pub mod code_handler;
 pub mod macros;
 pub mod preprocessor;